backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
motor>=3.0.0
pymongo>=4.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.2.0
numpy>=1.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

